#9758: Moved maxpool to pool namespace
