+++ ttnn/cpp/pybind11/operations/__init__.hpp
@@ -15,7 +15,6 @@
 #include "pybind11/operations/embedding.hpp"
 #include "pybind11/operations/kv_cache.hpp"
 #include "pybind11/operations/matmul.hpp"
-#include "pybind11/operations/maxpool2d.hpp"
 #include "pybind11/operations/normalization.hpp"
 #include "pybind11/operations/pool.hpp"
 #include "pybind11/operations/copy.hpp"
@@ -65,8 +64,8 @@ void py_module(py::module& module) {
     auto m_conv2d = module.def_submodule("conv2d", "conv2d operation");
     conv2d::py_module(m_conv2d);
 
-    auto m_maxpool2d = module.def_submodule("maxpool2d", "maxpool 2d operation");
-    maxpool2d::py_module(m_maxpool2d);
+    // auto m_maxpool2d = module.def_submodule("maxpool2d", "maxpool 2d operation");
+    // maxpool2d::py_module(m_maxpool2d);
 
     auto m_transformer = module.def_submodule("transformer", "transformer operations");
     transformer::py_module(m_transformer);
