+++ ttnn/cpp/pybind11/operations/pool.hpp
@@ -58,8 +58,37 @@ void bind_global_avg_pool2d(py::module& module) {
 }
 
 }  // namespace detail
+using array2_t = std::array<uint32_t, 2>;
 
-void py_module(py::module& module) { detail::bind_global_avg_pool2d(module); }
+void py_module(py::module& module) {
+    detail::bind_global_avg_pool2d(module);
+    module.def(
+        "maxpool2d",
+        [](const ttnn::Tensor& input_tensor,
+            uint32_t batch_size,
+            uint32_t input_height,
+            uint32_t input_width,
+            uint32_t channels,
+            array2_t kernel_size,
+            array2_t stride,
+            array2_t padding,
+            array2_t dilation,
+            Device& device) -> Tensor {
+            return maxpool2d(input_tensor, batch_size, input_height, input_width, channels, kernel_size, stride, padding, dilation, device);
+        },
+        py::kw_only(),
+        py::arg("input_tensor"),
+        py::arg("batch_size"),
+        py::arg("input_height"),
+        py::arg("input_width"),
+        py::arg("channels"),
+        py::arg("kernel_size"),
+        py::arg("stride"),
+        py::arg("padding"),
+        py::arg("dilation"),
+        py::arg("device")); }
+
+
 
 }  // namespace pool
 }  // namespace operations
